Import scripts for Feb 2025 E&O Open Data Release
These are the scripts and metadata needed for the 2025 Feb ATLAS
Education and Outreach Open Data Release. These are based on the recent
proton-proton Research Open Data release, but with several simplified
formats for various educational projects and skims that satisfy a
variety of use-cases. We'll need 25 new DOIs and record IDs for this
release, and we should plan for a news item as usual.

Please let me know if anything seems amiss, or if I've missed anything
in here!