Adding scripts for ATLAS HEPMC Open Data handling
This is the first iteration on scripts required for handling of the new
ATLAS event generation open data in HEPMC format. The README explains
all the different scripts and files included here (or should).

A few samples have been transferred to CERN already to establish the
functionality of all the scripts. Everything seems to be ok so far.

The key outstanding item (needed before anything can actually go onto
the Open Data Portal) is the record ID and DOI list for all the various
records that will be created. Otherwise this should be just about ready
to go, at least to the QA portal for checking.