Adding script for building js file from metadata csv
This js file then goes into the ATLAS open data website to document the
metadata. Including it here so that everything is in one place. Because
it is really just a copy of the csv with a header and footer, not trying
to include the output here as well (just unnecessary extra files)